Make Utils::Pi return float

+++ utils.h
@@ -73,7 +73,7 @@ public:
         }
     }
 
-    static constexpr double Pi() { return std::atan2(0, -1); }
+    static constexpr float Pi() { return std::atan2(0, -1); }
 };
 
 #endif // UTILS_H
